combine attachment.insert attachment.insertAsStream
