+++ lib/nano.js
@@ -444,7 +444,7 @@ module.exports = exports = function dbScope (cfg) {
   function followDb (dbName, qs0, callback0) {
     const { opts, callback } = getCallback(qs0, callback0)
     opts.db = urlResolveFix(cfg.url, encodeURIComponent(dbName))
-    // opts.httpAgent = httpAgent
+    opts.httpAgent = require('request')
     if (typeof callback === 'function') {
       return followAgent(opts, callback)
     } else {
@@ -869,19 +869,6 @@ module.exports = exports = function dbScope (cfg) {
       }, callback)
     }
 
-    function insertAttAsStream (docName, attName, att, contentType, opts) {
-      return relax({
-        db: dbName,
-        att: attName,
-        method: 'PUT',
-        contentType: contentType,
-        doc: docName,
-        qs: opts,
-        body: att,
-        dontStringify: true
-      })
-    }
-
     function getAtt (docName, attName, qs0, callback0) {
       const { opts, callback } = getCallback(qs0, callback0)
 
@@ -1094,7 +1081,6 @@ module.exports = exports = function dbScope (cfg) {
       },
       attachment: {
         insert: insertAtt,
-        insertAsStream: insertAttAsStream,
         get: getAtt,
         getAsStream: getAttAsStream,
         destroy: destroyAtt
